# docs/1.0.0-SNAPSHOT-api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
tags:
- name: offers
  description: offers
- name: accounts
  description: accounts
paths:
  /offers:
    get:
      tags:
      - offers
      - Offers
      summary: Get Offers
      operationId: getOffers
      parameters:
      - name: An Authorisation Token as per https://tools.ietf.org/html/rfc6750
        in: header
        required: true
        schema:
          type: string
      - name: "The time when the PSU last logged in with the TPP.  All dates in the\
          \ HTTP headers are represented as RFC 7231 Full Dates. An example is below:\
          \  Sun, 10 Sep 2017 19:43:31 UTC"
        in: header
        required: false
        schema:
          type: string
      - name: The PSU's IP address if the PSU is currently logged in with the TPP.
        in: header
        required: false
        schema:
          type: string
      - name: An RFC4122 UID used as a correlation id.
        in: header
        required: false
        schema:
          type: string
      - name: Indicates the user-agent that the PSU is using.
        in: header
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/OBReadOffer1"
            application/json:
              schema:
                $ref: "#/components/schemas/OBReadOffer1"
            application/jose+jwe:
              schema:
                $ref: "#/components/schemas/OBReadOffer1"
            application/json; charset=utf-8:
              schema:
                $ref: "#/components/schemas/OBReadOffer1"
  /accounts/{AccountId}/offers:
    get:
      tags:
      - Offers
      - accounts
      summary: Get Offers
      operationId: getAccountsAccountIdOffers
      parameters:
      - name: AccountId
        in: path
        required: true
        schema:
          type: string
      - name: An Authorisation Token as per https://tools.ietf.org/html/rfc6750
        in: header
        required: true
        schema:
          type: string
      - name: "The time when the PSU last logged in with the TPP.  All dates in the\
          \ HTTP headers are represented as RFC 7231 Full Dates. An example is below:\
          \  Sun, 10 Sep 2017 19:43:31 UTC"
        in: header
        required: false
        schema:
          type: string
      - name: The PSU's IP address if the PSU is currently logged in with the TPP.
        in: header
        required: false
        schema:
          type: string
      - name: An RFC4122 UID used as a correlation id.
        in: header
        required: false
        schema:
          type: string
      - name: Indicates the user-agent that the PSU is using.
        in: header
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/OBReadOffer1"
            application/json:
              schema:
                $ref: "#/components/schemas/OBReadOffer1"
            application/jose+jwe:
              schema:
                $ref: "#/components/schemas/OBReadOffer1"
            application/json; charset=utf-8:
              schema:
                $ref: "#/components/schemas/OBReadOffer1"
components:
  schemas:
    OBReadOffer1:
      type: object