# src/main/java/com/openbanking/offers/model/OBReadOffer1.java
package com.openbanking.offers.model;

import java.io.Serializable;
import lombok.Getter;
import lombok.Setter;

/**
 * OBReadOffer1
 */
@Getter
@Setter
public class OBReadOffer1 implements Serializable {
	private static final long serialVersionUID = 1L;

}